Update to current repo version

+++ scripts/deploy.sh
@@ -6,12 +6,17 @@ pull_latest() {
 }
 
 restart_app() {
-  # Attempt to restart unicorn gracefully as per
-  #  http://unicorn.bogomips.org/SIGNALS.html
-  pid=$(<"WcaOnRails/pids/unicorn.pid")
-  kill -SIGUSR2 $pid
-  sleep 5
-  kill -SIGQUIT $pid
+  if ps -efw | grep "unicorn master" | grep -v grep; then
+    # Found a unicorn master process, restart it gracefully as per
+    #  http://unicorn.bogomips.org/SIGNALS.html
+    pid=$(<"WcaOnRails/pids/unicorn.pid")
+    kill -SIGUSR2 $pid
+    sleep 5
+    kill -SIGQUIT $pid
+  else
+    # We could not find a unicorn master process running, lets start one up!
+    (cd WcaOnRails; bundle exec unicorn -D -c config/unicorn.rb)
+  fi
 }
 
 commit_hash() {
@@ -120,7 +125,18 @@ rebuild_regs() {
   rm -rf $tmp_dir
 }
 
+update_docs() {
+  documents_folder=WcaOnRails/public/documents
+  tmp_dir=/tmp/wca-documents-clone
+
+  rm -rf $tmp_dir
+  git clone --depth=1 --branch=build https://github.com/thewca/wca-documents.git $tmp_dir
+  rm -rf $documents_folder
+  mv $tmp_dir $documents_folder
+}
+
 restart_dj() {
+  sudo supervisorctl update
   sudo supervisorctl restart workers:*
 }
 
@@ -141,10 +157,13 @@ rebuild_rails() {
 cd "$(dirname "$0")"/..
 
 if [ "$(hostname)" == "production" ] || [ "$(hostname)" == "staging" ]; then
-  export RAILS_ENV=production
+  export RACK_ENV=production
 else
-  export RAILS_ENV=development
+  export RACK_ENV=development
 fi
 
-allowed_commands="pull_latest restart_app restart_dj rebuild_rails rebuild_regs"
+# Workaround for https://github.com/rails/webpacker/issues/773
+export RAILS_ENV=${RACK_ENV}
+
+allowed_commands="pull_latest restart_app restart_dj rebuild_rails rebuild_regs update_docs"
 source scripts/_parse_args.sh
